backend/requirements.txt
absl-py==2.4.0
av==16.1.0
contourpy==1.3.3
cycler==0.12.1
einops==0.8.2
fastapi==0.115.0
flatbuffers==25.12.19
fonttools==4.61.1
h5py==3.15.1
heartpy==1.2.7
jax==0.9.0.1
jaxlib==0.9.0.1
keras==3.13.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
numpy==2.4.2
onnxruntime==1.24.2
open-rppg==0.1.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
rich==14.3.3
scipy==1.17.0
setuptools==69.5.1
six==1.17.0
sympy==1.14.0
typing_extensions==4.15.0
uvicorn==0.30.0
wheel==0.46.3
